Add basic NDK pieces.
These get libc++ building for the NDK, and the majority of tests
passing for armeabi-v7a on GB, ICS, JB, and L.

This is a trivial update to r263688 (using danielaustin's platform
update that had to be reverted, but it doesn't have issues for the
NDK). There's more that can be pruned down (we probably don't need
src/support/android/locale_android.cpp any more), but I'd rather have
the trivial update submitted separately so we can more easily track
the changes.

This still differs from upstream slightly. We still have
https://android-review.googlesource.com/#/c/181691/ because it has
caused trouble when building things with long doubles for x86_64
(this might now be resolved; someone should check). We also still have
https://android-review.googlesource.com/#/c/102419/. It looks like
some other things have been changed in this area for other platforms,
so this might be something to either upstream or start using the
things that upstream uses (will need some investigation).

Bug: http://b/19149083
Bug: https://github.com/android-ndk/ndk/issues/106
Change-Id: I36106f4af9b9dee9521300ae16d4fdc5d0f4b52f