Export libc++abi includes for the shared libc++.
These are automatically available in the static version because apps
using the static STL already have to link libc++abi.a.

For some reason our build time checks aren't building the test that
caught this (asan-smoke). Probably has something to do with the fact
that this test can only be run on very new platform levels. I need to
look in to that.

Change-Id: I7817e39aafbe17a5156f26b8843077b7aad7eb61