Add a module for the prebuilt libunwind.
We've thus far been missing this in r13. Apparently ndk-build doesn't
actually consider non-existent dependencies to be an error (or even a
warning), so the lack of this module hasn't caused any build time
issues.

I'm adding this to libc++'s makefile rather than libunwind because we
don't want to expose libunwind as an NDK API.

Test: ndk/run_tests.py --filter libcxx-unwind-static
Bug: http://b.android.com/223708
Change-Id: I58e8c6787e3e282556bb3a4f0267998fa5eeee70