Fix the NDK test runner.
This apparently got horribly broken in a refactoring at some point,
and since running on older devices takes *ages*, I guess it hasn't
been run on anything pre-L in a while.

* Actually links libandroid_support. I think this was a bad
  refactoring while playing with using the linker script here.
* Detect the device API level to determine PIE setting. This was
  previously just manually fiddled with.
* Use the correct API level's headers. It looks like the
  autogenerated file and the input file got mixed at some point :\

The good news is that there are only a handful of new failures when
running on Jelly Bean (targeting Gingerbread) with these fixed. The
new failures seem to all be because `_Exit` was added to the
libandroid_support headers but was never actually implemented.

Test: ndk/test_libcxx.py -a armeabi-v7a -p 9 # on Nakasi android-16
Bug: None
Change-Id: I5d81ada9c0a4cec6eee18be8a98996c89265198f