Disable XFAIL handling for the NDK tests.
The libc++ test runner doesn't provide the level of detail we need to
handle XFAILs in the NDK test runner; the NDK distinguishes between
failed builds and failed runs because we have a --skip-run option.

Clear all the XFAIL data in the test before the test runner has a
chance to check against it. We'll handle marking expected failures in
the usual test_config.py manner.

Test: ndk/test_libcxx.py --abi armeabi-v7a
Bug: None
Change-Id: I39078068f6fc6c5093a022492f50099ad7b310be