Reset globalMemCounter before using it.
If any static constructors call new or delete, these counters won't
be at zero when we enter the test. The concrete example of this
happens when building the tests with the NDK, where
libandroid_support (a compatibility library that allows libc++ to
work on pre-O devices) calls new in its implementation of newlocale,
which gets called as part of iostreams initialization.

Test: ndk/run_tests.py
Bug: None
Change-Id: Ic08854685dd9714ba02c50a2f6cc3828e2f7577c