Get test source from the source directory.
Clean up our config so we can use the libc++ tests from the source
directory rather than the NDK install.

Remove the checked in test/lit.site.cfg (used by the platform tests,
not the NDK) and add the file to .gitignore so we don't make a mess of
the source directory.

Test: ndk/run_tests.py --rebuild
Bug: None
Change-Id: I82af04c623d2e5cf7b635506ff7770d0a5e68bad